Rename daysSequenceStorageTests to daysChallengeStorageTests
